Rename ClientInfo to PipeClient.
